GRID #1 MERGE

+++ grid/src/main/java/org/highfives/grid/performance_review/query/dto/ReviewItemDTO.java
@@ -16,8 +16,8 @@ public class ReviewItemDTO {
     private int weight;
     private String performance;
     private int selfId;
-    private float selfScore;
+    private double selfScore;
     private String selfComment;
     private int superiorId;
-    private float superiorScore;
+    private double superiorScore;
 }
